Add CardModel
